fix: escape body of post requests
The body of a request of type application/x-www-form-urlencoded must be escaped. Currently the data is encoded to base64, which will usually be valid has the change of producing +, /, or = characters. If left unescaped, this will not be handled properly.

See: https://www.w3.org/TR/html401/interact/forms.html#h-17.13.4.1